requirements.txt
tensorflow-gpu
keras
pandas
matplotlib
numpy
sklearn
pillow
opencv-python
